+++ src/components/SpStandardIconText.vue
@@ -1,0 +1,30 @@
+<template>
+  <IconText
+    class="standard-icon-text"
+    :text="text"
+    :link="link"
+    :iconType="iconType"
+    :isIconFirst="true"
+  />
+</template>
+
+<script>
+import IconText from './texts/IconText'
+
+export default {
+  components: {
+    IconText
+  },
+  props: {
+    text: { type: String, required: true },    
+    link: { type: String, default: null },    
+    iconType: {
+      type: String,
+      default: 'check',
+      validator: function(value) {
+        return ['copy', 'check', 'successDot'].indexOf(value) !== -1
+      }               
+    }
+  }
+}
+</script>
