+++ src/components/blocks/BlockChain.vue
@@ -48,10 +48,10 @@
 import { mapGetters, mapMutations, mapActions } from 'vuex'
 import moment from 'moment'
 
-import { formatter as blockFormatter } from '../../../helpers/block'
+import { formatter as blockFormatter } from '../../helpers/block'
 
-import BlockCard from '../../cards/BlockCard'
-import IconArrow from '../../../assets/icons/Arrow'
+import BlockCard from '../cards/BlockCard'
+import IconArrow from '../../assets/icons/Arrow'
 
 export default {
   props: {
